refactoring : array를 객체로 승격
